Add a new Key-Value transaction type and a test for DiskDb.
I have added a new Key-Value pair called `KeyValue::BlockHeight`,
it maps the Monero block IDs (hashes) to the height they are in.

Also I've added a test for `DiskDb` that tests the insertion
of `KeyValue::BlockHeight` into the db. This test should be expanded
to cover more code, also the `fn write` and `fn get` methods of
`DiskDb` need to be cleaned a little bit to make them more readable
and less error prone.

I've modified the `serialization` crate to implement `Serialize`
and `Deserialize` for integer types.

The dxmr main file was fixed to give a directory and not a file
for `DiskDb`.

Signed-off-by: Jean Pierre Dudey <jeandudey@hotmail.com>